+++ CS2-Clase-1/src/main/java/app/domain/models/Partner.java
