+++ CS2-Clase-1/src/main/java/app/domain/models/Person.java
@@ -1,18 +1,26 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
+package app.domain;
 
-package app.domain.models;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 
-/**
- *
- * @author emman
- */
+@Setter
+@Getter
+@NoArgsConstructor
 public class Person {
-    private String name;
-    private int age;
+    
+    private long personId;
     private long document;
-    
-    
+    private String name;  
+    private long cellPhone;
+
+    public Person(long personId, long document,String name,  long cellPhone) {
+
+        this.personId = personId;
+        this.document = document;
+        this.name = name;
+        this.cellPhone = cellPhone;
+    }
+
+     
 }
