+++ CS2-Clase-1/src/main/java/app/domain/services/GuestServices.java
@@ -1,0 +1,13 @@
+package app.domain.services;
+
+import app.domain.models.Guest;
+
+public class GuestServices {
+
+    private PersonPort personPort;
+
+    public void makeConsuption(Guest gest){
+        
+    }
+}
+
