+++ CS2-Clase-1/src/main/java/app/ports/PersonPort.java
@@ -1,0 +1,5 @@
+package app.ports;
+
+public class PersonPort {
+
+}
